Tidy up the code created for 'Single item - join, bid & lose'
TDD Notes: Once the end to end tests are passing, its a best practice to
come back to refactor and clean the code that was created under the
pressure of passing the test cases.
If testing and production code are properly isolated this step shouldn't
trigger changes in the testing code.