Show price details
Modify the user interface to show the price details of an auction.

TDD Notes: functionality is added in little slices making sure that
everything remains working step after step, only end to end test cases
could be breaking for longer time as indicators of a feature
implementation that is still in progress.