Add new items through the GUI
TDD Notes: integration test cases are used instead of a unit test case
when it's not possible to isolate the code under test from a framework.
For example, the only way to test a simple interaction with the GUI
(which is the unit in this case) is by starting up the GUI through the
Swing framework.